Remove flashing text with the player name from the Next track label in WPS
Currently, in Cabbie v2 there is a demonstration of the %ft() tag, which shows the player name (or Rockbox! if not configured) for a second between switching tracks, or every 5 minutes of playback. This is confusing for a couple of users (including me), does not bring any useful information, and is removed from the theme. The tag and its documentation are left intact, so theme developers can make use of it in their themes.

Change-Id: Id3c79f442e4836a78d24997bd66e01c08206dd18