rgnano:  More aggressively update the "installed" version
The current code checks the date listed in the "Version" string
in rockbox-info.txt, and updates if the new package has a newer date.

However this is not granular enough, especially when you are iterating
multiple builds on a single day.  It also prevents downgrading to
an older build.

So, instead of checking for a newer date, apply the update if the new
version string (including the commit id) is merely *different* than the
installed one.

Change-Id: Ie5d10ffe011afd1efb3c76c77a3ed19d36df9321