firmware: fix misdefinition of QUEUE_GET_THREAD macro
This tries to access 'struct event_queue->send', but that
is only available if HAVE_EXTENDED_MESSAGING_AND_NAME is set.
This is not defined for bootloaders; this is a problem when
trying to build a bootloader with debugging enabled, because
the misdefined macro is used in some KERNEL_ASSERTs that get
compiled out on non-debug builds.

Change-Id: I334eedcda1ee7047c8dddcb7fa0c9717156f2a0a